🔨 refac(user/service) Interface 분리  #96
