+++ src/main/java/Backend/HIFI/domain/user/service/UserServiceImpl.java
@@ -1,0 +1,50 @@
+package Backend.HIFI.domain.user.service;
+
+import Backend.HIFI.domain.user.entity.User;
+import Backend.HIFI.domain.user.repository.UserRepository;
+import Backend.HIFI.domain.follow.repository.FollowRepository;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class UserServiceImpl implements UserService {
+
+    private final UserRepository userRepository;
+    private final FollowRepository followRepository;
+
+    @Override
+    public User findByEmail(String email) {
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new IllegalArgumentException("가입되지 않은 Email 입니다"));
+        return user;
+    }
+
+    @Override
+    public User findById(Long id) {
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("가입되지 않은 ID 입니다"));
+        return user;
+    }
+
+    private void validateDuplicateUser(User user) {
+        User findUser = userRepository.findById(user.getId())
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 유저입니다"));
+        if (findUser == user) {
+            throw new IllegalStateException("이미 존재하는 회원입니다");
+        }
+    }
+
+    @Override
+    public void deleteUser(User user) {
+        // trigger 통해 리뷰 지워지면 store.review 지워지도록
+//        userRepository.deleteReviewByUserId(user.getId());
+        userRepository.deleteFollowByUserId(user.getId());
+//        userRepository.delete(user);
+        user.updateIsDeleted();
+    }
+
+}
